Sample 24bit

+++ CM7/Core/Inc/fft_hist.h
@@ -3,7 +3,7 @@
 
 #include "stdint.h"
 
-void fft_16hist(int16_t *out_l, int16_t *out_r, int16_t *in);
-void fft_16p(float *out, int16_t *in);
+void fft_16hist(int16_t *out_l, int16_t *out_r, int32_t *in);
+void fft_16p(float *out, int32_t *in);
 
 #endif /* FFT_HIST_H */
